Tasks

+++ src/main/java/tech/mms/cos/api/SpringController.java
@@ -1,39 +1,46 @@
 package tech.mms.cos.api;
 
-import org.springframework.cglib.core.Local;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import tech.mms.cos.api.client.EmployeesApi;
 import tech.mms.cos.api.mapper.EmployeeApiMapper;
 import tech.mms.cos.api.model.EmployeeDTO;
 import tech.mms.cos.api.model.NewEmployeeRequestDTO;
+import tech.mms.cos.core.AuthenticationService;
 import tech.mms.cos.core.RandomEmployeeService;
-import tech.mms.cos.core.model.Employee;
-import tech.mms.cos.core.model.Genders;
-import tech.mms.cos.core.model.Name;
+import tech.mms.cos.core.model.*;
 import tech.mms.cos.exception.AppValidationException;
 import tech.mms.cos.repository.EmployeeRepository;
+import tech.mms.cos.service.GitHubAuthService;
+import tech.mms.cos.service.GitHubOAuthRequest;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.stream.Collectors;
 
 import static tech.mms.cos.core.model.Genders.M;
 
 @RestController
+@RequestMapping("/")
 public class SpringController implements EmployeesApi {
 
     private final EmployeeRepository employeeRepository;
     private final RandomEmployeeService employeeGeneratorApi;
+    private final GitHubAuthService gitHubAuthService;
 
-    public SpringController(EmployeeRepository employeeRepository, RandomEmployeeService employeeGeneratorApi) {
+    public SpringController(EmployeeRepository employeeRepository, RandomEmployeeService employeeGeneratorApi, GitHubAuthService gitHubAuthService) {
         this.employeeRepository = employeeRepository;
         this.employeeGeneratorApi = employeeGeneratorApi;
+        this.gitHubAuthService = gitHubAuthService;
     }
 
 
     @GetMapping("/getDummy")
+    @PreAuthorize("hasAnyRole('EDITOR', 'ADMIN')")
     public Employee getDummy() {
         var employee = new Employee(LocalDate.of(1990, 1, 1),
                 20, 40, M, new Name("Test", "Dummy", "Unit"), "HR");
@@ -51,10 +58,16 @@ public class SpringController implements EmployeesApi {
         // 4. Schau dir die Response der API an, baue eine Klasse (Mehrere Klassen) die der Response entsprechen.
         // https://randomuser.me/api/
 
+
+
         var randomEmployee = employeeGeneratorApi.generate();
         employeeRepository.saveEmployee(randomEmployee);
         return randomEmployee;
     }
+
+
+
+
 
 
     @Override
@@ -63,9 +76,71 @@ public class SpringController implements EmployeesApi {
         return ResponseEntity.ok(EmployeeApiMapper.mapResponse(employee));
     }
 
+
     @Override
     public ResponseEntity<Void> employeesUuidDelete(String uuid) {
+
+    /*
+
+            3 Rollen für Accounts: Viewer, Editor, Admin
+            Sinnvolle authorization für die jeweiligen Eployee Endpoints, z.B. getAll -> Viewer&Editor, delete -> Editor, Admin -> Alles
+
+            Wenn Rolle < ADMIN, dann braucht Account für Löschen das gleiche Department wie Employee
+
+             */
+
+        var employee = employeeRepository.find(UUID.fromString(uuid))
+                .orElse(null);
+
+        if(employee == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        // Was ist OAuth 2.0, was ist ein JWT Token, Aufbau JWT Token, Was sind OAuth Flows, Authorization code flow (Video)
+        // Video: How to use OAtuh with Github (Anlegen einer Github OAuth App)
+
+        /*
+
+        ##-- OAuth Flows --##
+        1. Baue im UI einen Login Button, welcher dich direkt auf Github weiterleitet zum Authorisieren
+        Client ID -> Kann ich öffentlich machen
+        Client Secret -> Nur im Backend als Secret, (nicht öffentlich machen)
+
+
+        2. Bauene eine Page auf der Redirect Url,
+         welche den "code" einließt und zum Backend schickt auf {backendUrl}/authlogin/oauth/github/access_token
+
+        3. Erstellen im Backend einen AuthController,
+         welcher auf /auth/login/oauth/github/access_token hört,
+          den "code" entgegennimmt, und den Code für ein OAuth Token austauscht
+
+        4. Das Frontend kriegt den OAuth Token vom Backened
+        als antwort zurück und speichert das OAuth Token im Redux Store
+
+
+        https://github.com/settings/applications/3045564
+        https://docs.github.com/de/apps/oauth-apps/building-oauth-apps/authorizing-oauth-apps
+        https://docs.github.com/de/apps/oauth-apps/building-oauth-apps/scopes-for-oauth-apps
+
+
+         */
+
+        AuthenticatedAccount auth = AuthenticationService.getAccountAuth();
+
+        if (!auth.hasHigherPermissionThan(AccountRole.EDITOR)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        if (auth.isBetween(AccountRole.EDITOR, AccountRole.ADMIN)) {
+            var userDepartments = auth.getDetails().getDepartments();
+            if (!userDepartments.contains(employee.getDepartment())) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+        }
+
         var wasDeleted = employeeRepository.deleteEmployee(UUID.fromString(uuid));
+
+        SecurityContextHolder.getContext().getAuthentication();
 
         if (wasDeleted) {
             return ResponseEntity.noContent().build();
@@ -105,8 +180,9 @@ public class SpringController implements EmployeesApi {
 
             LocalDate birthdate = LocalDate.parse(employeeDTO.getBirthdate());
 
-            existingEmployee.setBirthdate(birthdate);
-            existingEmployee.setHourlyRate(employeeDTO.getHourlyRate());
+            existingEmployee
+                    .setBirthdate(birthdate)
+                    .setHourlyRate(employeeDTO.getHourlyRate());
             existingEmployee.setHoursPerWeek(employeeDTO.getHoursPerWeek());
             existingEmployee.setGender(Genders.valueOf(employeeDTO.getGender().name()));
             existingEmployee.setName(new Name(
@@ -136,7 +212,6 @@ public class SpringController implements EmployeesApi {
                 .toList()
         );
     }
-
 
 }
